devdox/pyproject.toml
[tool.setuptools]
packages = ["app"]

[project]
name = "context-creators"
version = "0.1.1"
description = "Context Creators Application"
requires-python = ">=3.12"
dependencies = [
    # Core dependencies
    "annotated-types==0.7.0",
    "aiohttp==3.11.18",
    "anyio==4.9.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    
    # Tortoise ORM and database
    "tortoise-orm>=0.20.0",
    "asyncpg>=0.28.0",
    "aerich>=0.7.2",
    
    # FastAPI and web framework
    "fastapi==0.115.12",
    "fastapi-cli==0.0.7",
    "starlette==0.46.2",
    "uvicorn==0.34.2",
    "uvloop==0.21.0",
    
    # HTTP and networking
    "charset-normalizer==3.4.2",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "urllib3==2.4.0",
    "websockets==14.2",
    "watchfiles==1.0.5",
    
    # Data validation and parsing
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
    "pydantic_core==2.33.2",
    "email_validator==2.2.0",
    
    # Utilities
    "click==8.2.0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "PyYAML==6.0.2",
    "typer==0.15.3",
    "typing-inspection==0.4.0",
    "typing_extensions==4.13.2",
    "sniffio==1.3.1",
    
    # Security and crypto
    "cryptography==44.0.3",
    "PyJWT<3.0.0,>=2.6.0",
    "clerk-backend-api==2.2.0",
    
    # Templates and rendering
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.2",
    "markdown-it-py==3.0.0",
    "mdurl==0.1.2",
    
    # Rich text and CLI
    "rich==14.0.0",
    "rich-toolkit==0.14.6",
    "Pygments==2.19.1",
    "shellingham==1.5.4",
    
    # DNS and networking
    "dnspython==2.7.0",
    
    # Git and version control
    "python-gitlab==5.6.0",
    "GitPython==3.1.44",
    "PyGithub==2.6.1",
    "requests==2.32.4",
    
    # Real-time and webhooks
    "realtime==2.4.3",
    "svix==1.66.0",
    
    # Utilities
    "pycparser==2.22",
    "tomli-w==1.2.0",

    # Tests
    "pytest-cov==6.1.1",
    "pytest-asyncio==1.0.0",
    "pytest-tornasync==0.6.0.post2",
    "pytest-trio==0.8.0",
    "pytest-twisted==1.14.3",
    "twisted>=22.0.0",
    "coverage==7.8.0",
]

[tool.poetry.extras]
dev = ["pytest", "coverage"]


[tool.aerich]
tortoise_orm = "app.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."